+++ pkg/types/namespace.go
@@ -24,7 +24,7 @@ package types
 
 // NamespaceItem represents a namespace.
 type NamespaceItem struct {
-	ID          uint    `json:"id"`
+	ID          uint64  `json:"id"`
 	Name        string  `json:"name" validate:"required,min=2,max=20,is_valid_namespace"`
 	Description *string `json:"description" validate:"max=30"`
 
@@ -50,17 +50,17 @@ type ListNamespaceRequest struct {
 
 // GetNamespaceRequest represents the request to get a namespace.
 type GetNamespaceRequest struct {
-	ID uint `json:"id" param:"id" validate:"required,number"`
+	ID uint64 `json:"id" param:"id" validate:"required,number"`
 }
 
 // DeleteNamespaceRequest represents the request to delete a namespace.
 type DeleteNamespaceRequest struct {
-	ID uint `json:"id" param:"id" validate:"required,number"`
+	ID uint64 `json:"id" param:"id" validate:"required,number"`
 }
 
 // PutNamespaceRequest represents the request to update a namespace.
 type PutNamespaceRequest struct {
-	ID uint `json:"id" param:"id" validate:"required,number"`
+	ID uint64 `json:"id" param:"id" validate:"required,number"`
 
 	Description *string `json:"description" validate:"max=30"`
