+++ pkg/types/tag.go
@@ -24,7 +24,7 @@ package types
 
 // TagItem represents an tag.
 type TagItem struct {
-	ID        uint   `json:"id"`
+	ID        uint64 `json:"id"`
 	Name      string `json:"name"`
 	Digest    string `json:"digest"`
 	Size      int64  `json:"size"`
@@ -42,14 +42,14 @@ type ListTagRequest struct {
 
 // DeleteTagRequest represents the request to delete a tag.
 type DeleteTagRequest struct {
-	ID         uint   `json:"id" param:"id" validate:"required,number"`
+	ID         uint64 `json:"id" param:"id" validate:"required,number"`
 	Namespace  string `json:"namespace" param:"namespace" validate:"required,min=2,max=20,is_valid_namespace"`
 	Repository string `json:"repository" query:"repository" validate:"required,is_valid_repository"`
 }
 
 // GetTagRequest represents the request to get a tag.
 type GetTagRequest struct {
-	ID         uint   `json:"id" param:"id" validate:"required,number"`
+	ID         uint64 `json:"id" param:"id" validate:"required,number"`
 	Namespace  string `json:"namespace" param:"namespace" validate:"required,min=2,max=20,is_valid_namespace"`
 	Repository string `json:"repository" query:"repository" validate:"required,is_valid_repository"`
 }
